apply solid principles

+++ librarysystem/src/main/java/com/librarysystem/book/service/BookService.java
@@ -69,8 +69,8 @@ public class BookService {
     }
 
     public void deleteBook(Long id) {
-        Books book = findBookById(id)
-                .orElseThrow(() -> new NoSuchElementException("Book not found."));
+        if(!bookRepository.existsById(id))
+                throw new NoSuchElementException("Book not found.");
 
         bookRepository.deleteById(id);
     }
